Aimlessly writing code

+++ src/BudgetTracker.Shared/Domain.fs
@@ -2,8 +2,10 @@
 
 open System
 
+// TODO: Decimal comparison seems to be broken in Fable, create repro and file an issue.
+
 type Expense = {
-  Amount: decimal
+  Amount: float
   Reason: string
   OccuredAt: DateTime
 }
@@ -12,7 +14,7 @@ type AvailableCategory = { Name: string }
 
 type AllocatedCategory = {
   Name: string
-  Allocation: decimal
+  Allocation: float
   Expenses: Expense list
 }
 
@@ -25,6 +27,6 @@ type Budget = {
   OwnerId: string
   Name: string
   Description: string option
-  EstimatedMonthlyIncome: decimal
+  EstimatedMonthlyIncome: float
   Categories: Category list
 }
